+++ implementacao/src/config.json
@@ -1,20 +1,20 @@
 {
-    "tcp_port": 1234,
+    "server": {
+        "send_interval": 30,
+        "url": "http://localhost:5000"
+    },
     "i2c_sensors": {
         "distance_sensor_addr": 38,
         "battery_sensor_addr": 97,
         "bus": 100,
         "light_sensor_addr": 16
     },
-    "server": {
-        "url": "http://localhost:5000/test",
-        "send_interval": 30
-    },
     "gprs": {
-        "apn": "tim.com.br",
         "port": "/dev/ttyACM200",
+        "apn": "tim.com.br",
         "rate": 9600
     },
+    "tcp_port": 1234,
     "gps": {
         "port": "/dev/ttyACM100",
         "rate": 9600
